Сохранение изменений (рабочие)

+++ src/main/java/ru/skyeng/postservice/repository/PostOfficeRepository.java
@@ -6,7 +6,7 @@ import ru.skyeng.postservice.model.PostOffice;
 
 import java.util.Optional;
 
-public interface PostOfficeRepository extends JpaRepository<PostOffice, Integer> {
+public interface PostOfficeRepository extends JpaRepository<PostOffice, Long> {
 
     @Query("select p from PostOffice p " +
             "where p.index=?1")
