+++ src/main/java/ru/skyeng/postservice/repository/StageDeliveryRepository.java
@@ -3,6 +3,6 @@ package ru.skyeng.postservice.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.skyeng.postservice.model.StageDelivery;
 
-public interface StageDeliveryRepository extends JpaRepository<StageDelivery, Integer> {
+public interface StageDeliveryRepository extends JpaRepository<StageDelivery, Long> {
 
 }
